Basically working apk-DAG building using merge+diff.
The code is very rough still, just a basic proof-of-concept.

Also, this is based on an older commit because on newer commits where I
have to use "dagger do" I am getting "FS not set" errors again. Not
sure, but suspect it could be related to the issue around "Dynamic
Tasks" mentioned in this recent PR:
https://github.com/dagger/dagger/pull/1691

This doesn't actually use apko right now and it seems like it may not
make sense to use apko until more features are upstreamed. Instead, this
just does the apk package sorting itself and then uses merge+diff to
combine them together efficiently.

As an example of what apko is currently missing: it always generates a
single-layer image, whereas this implementation generates an image with
one-layer-per-package, which has numerous caching benefits (thanks to
merge+diff). I am investigating upstreaming multilayered images to apko
and whether that would make it easier to re-use apko in the context of
Dagger/Buildkit or not.

In general though, we need to figure out if we want this to be an
"apko" package that just literally calls apko or if we want this to
be something that is similar in spirit to apko but distinct. There's
many tradeoffs and options here that are worth exploring more.

Some other significant outstanding issues that need to be addressed
(besides more mundane cleanup):
* Robust parsing of package dependencies (it's hacky cli output parsing
  right now)
* Support for version pinning
* When not using version pinning, vertexes should re-run if the latest
  version of a package has changed.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>